feat:4.21 delete Method refactored

+++ part4/blog/controllers/blogs.js
@@ -36,8 +36,20 @@ blogsRouter.post("/", async (req, res, next) => {
 
 blogsRouter.delete("/:id", async (req, res, next) => {
   try {
-    await Blog.findByIdAndRemove(req.params.id);
-    res.status(204).end();
+    const decodedToken = jwt.verify(req.token, process.env.SECRET);
+    if (!decodedToken.id) {
+      return res.status(401).json({ error: "token invalid" });
+    }
+
+    const user = await User.findById(decodedToken.id);
+    const blog = await Blog.findById(req.params.id);
+
+    if (blog.user.toString() === user.id.toString()) {
+      await Blog.findByIdAndRemove(req.params.id);
+      res.status(204).send("Blog deleted");
+    } else {
+      res.status(401).send("Unauthorized deletion tried");
+    }
   } catch (error) {
     next(error);
   }
